feat: 更新数据类型和添加批量添加域名功能
- 添加批量添加域名的功能。
- 完成域名总数量的显示。
- 优化日志记录和错误处理逻辑。
- 修改自增主键字段，将 `id` 和 `account_id` 字段从 `i32` 更新为 `i64`。
- 更新相关测试用例以适应新的数据类型和功能。

+++ crates/domain_manager/src/models/account.rs
@@ -5,7 +5,7 @@ use secrecy::SecretString;
 /// 账户模型
 #[derive(Debug, Clone)]
 pub struct Account {
-    pub id: i32,
+    pub id: i64,
     pub username: String,
     pub email: String,
     pub salt: String,
