+++ crates/domain_manager/src/storage/entities/domain.rs
@@ -5,9 +5,9 @@ use serde::{Deserialize, Serialize};
 #[sea_orm(table_name = "domains")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
-    pub id: i32,
+    pub id: i64,
     pub name: String,
-    pub provider_id: i32,
+    pub provider_id: i64,
     pub status: String,
     pub created_at: DateTime,
     pub updated_at: Option<DateTime>,
